Added Simple_onTrigger script
Fixed simple shoot to use a particle system that is on the GameObject.
Added Simple_onTrigger. It has 3 sections where events will be called
when something enters, exits, or stays in the trigger area.
Added a horrrid grass texture.
Added a song by Rolemusic.
Livened up the main scene with the new trigger system.